+++ build_runner_core/lib/build_runner_core.dart
@@ -7,7 +7,6 @@ export 'package:build/build.dart' show PostProcessBuildStep, PostProcessBuilder;
 export 'src/asset/batch.dart' show wrapInBatch;
 export 'src/asset/file_based.dart';
 export 'src/asset/finalized_reader.dart';
-export 'src/asset/reader.dart' show RunnerAssetReader;
 export 'src/asset/writer.dart';
 export 'src/environment/build_environment.dart';
 export 'src/environment/io_environment.dart';