+++ build_runner_core/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Start using `package:build/src/internal.dart'.
 - Use `build_test` 3.0.0.
 - Refactor `PathProvidingAssetReader` to `AssetPathProvider`.
+- Refactor `MultiPackageAssetReader` to internal `AssetFinder`.
 
 ## 8.0.0